+++ packages/ui/src/image-uploader/index.ts
@@ -1,0 +1,1 @@
+export * from './image-uploader';
