+++ packages/ui/src/index.ts
@@ -1,4 +1,3 @@
-export { default as Uploader } from './image-uploader';
 export { default as Skeleton } from './skeleton';
 export { default as Tabs } from './tabs';
 export { default as Tooltip } from './tooltip';
@@ -8,8 +7,9 @@ export * from './button';
 export * from './calendar';
 export * from './dialog';
 export * from './form';
-export * from './input';
 export * from './icon';
+export * from './image-uploader';
+export * from './input';
 export * from './popover';
 export * from './select';
 export * from './spinner';
