+++ packages/ui/src/toast/toaster.tsx
@@ -11,7 +11,7 @@ import {
 } from './toast';
 import { useToast } from './useToast';
 
-export function Toaster() {
+export const Toaster = () => {
   const { toasts } = useToast();
 
   return (
@@ -33,4 +33,4 @@ export function Toaster() {
       <ToastViewport />
     </ToastProvider>
   );
-}
+};
